Change slugs to use underscore (#83)
* Change slugs to use underscore

Update how handle is generated when creating an attribute. This is to prevent issues when parsing the attributes in javascript without adding additional mapping using php.
case : suppose we want to add an attribute named "warranty information", the handle generated will be 'warranty-information'. this commit attempts to fix this by replacing '-' with '_'

* Update AbstractAttribute.php

* Force slugs to be snake case instead of kebab case

* comply with setup standards

* Add a string function to correctly create handles

* Fix CI styles

* CI fixes

* CI fix

* CI fix

* moved handle generator to core && added unit tests

* Fix missing Str namespace

* fix CI style

Co-authored-by: itcyborg <isaac.itcyborg@outlook.com>