Hotfix - Properly handle adding media when disk is not local (#701)
* Properly handle adding media when disk is not local

In order to obtain the media from a remote location, addMediaFromDisk has to be used. addMedia only works when the file is stored on the server. You can set a default media disk for Spatie Media Library, so I used that config and then determined if it was local or not, so the image can be found and added properly.

* Added storage facade

* Pint fixes

* Pint

* Updated variables to camelcase

* Update HasImages.php

Co-authored-by: Alec Ritson <hello@itsalec.co.uk>