Feat - Refactor Order creation (#1097)
This PR aims to refactor and improve Orders and how they relate to
Carts. Currently a Cart can only have 1 order associated via the
`order_id` on the `carts` table, which was fine for most cases but it's
a bit restrictive, also the `CreateOrder` action has all the logic
within, which means if you want to customise any aspect of this, you
have to replace the action entirely, which isn't ideal if you only want
to do a small tweak...

The changes listed below will likely be breaking changes so this will
likely go in the next `0.X` release.


## Changes to the Cart > Order relationship.

Currently a cart has an `order_id` column, this will change so that the
`order` table now has the `cart_id` column, this will then also
introduce the concept of "draft" and "complete" orders.

### Draft Orders

These are orders which have been created from a cart that are yet to be
considered "paid" for.

### Complete Orders

These are orders which are considered "paid"

The logic which determines whether an order is complete is the presence
of a date in the `placed_at` column.

### Accessing the relationship

Previously.

```php
$cart->order
```

```php
// This is relationship which will accept an ID of the draft you want to fetch.
$cart->draftOrder
$cart->draftOrder(1);
```

```php
// Return all completed orders.
$cart->completedOrders

// Return the first completed order.
$cart->completedOrder

// Return the completed order by it's id
$cart->completedOrder(1);
```

It's also possible to check whether the cart has completed orders with a
little helper

```php
$cart->hasCompletedOrders();
```

## CreateOrder action changes.

The biggest change to this Action is that currently it's very much
"create an order every time you call me", which is not very useful, so
now the action will handle whether there is an existing order we can
just simply update. **Note, this action will specifically look for a
draft order to update, completed orders will not be touched**. The
process is as follows:

- Find a **draft** order if given an ID
- If no ID is passed or not found, return an instance of a new order to
create.
- If the cart already has completed order associated, check whether to
allow for multiple orders.
- If multiple orders are not allowed, throw an
`DisallowMultipleCartOrdersException`
- If the above passes, carry on.

As mentioned at the start, the logic for this is currently all within
the `CreateOrder` action class, which is not in keeping with the
direction we want to go in terms of offering flexibility for developers.
This has now be refactored into pipelines, much like when calculating
the cart totals.

These pipelines are stored in `config/lunar/orders.php` and the default
values will be:

```php
'pipelines' => [
    'creation' => [
        Lunar\Pipelines\Order\Creation\FillOrderFromCart::class,
        Lunar\Pipelines\Order\Creation\CreateOrderLines::class,
        Lunar\Pipelines\Order\Creation\CreateOrderAddresses::class,
        Lunar\Pipelines\Order\Creation\CreateShippingLine::class,
        Lunar\Pipelines\Order\Creation\CleanUpOrderLines::class,
        Lunar\Pipelines\Order\Creation\MapDiscountBreakdown::class,
    ],
],
```

An example of what these classes could look like is:

```php
<?php

namespace App\Pipelines\Orders\Creation;

use Closure;
use Lunar\Models\Order;
use Lunar\Models\OrderAddress;

class AcmeOrderPipeline
{
    /**
     * @return Closure
     */
    public function handle(Order $order, Closure $next)
    {
        // ...
        return $next($order);
    }
}

```

In addition to this, there are some changes to the `CreateOrder` action,
based on the draft/complete concept above.

Current execute method:

```php
public function execute(
    Cart $cart
)
```

New execute method:

```php
public function execute(
    Cart $cart,
    bool $allowMultipleOrders = false,
    int $orderIdToUpdate = null
): self
```
- `$allowMultipleOrders` - Determine whether you want to allow another
draft order to be created, if there is already a complete order
associated to this cart.
- `$orderIdToUpdate` - Pass through the ID of the order you want to
update.

## Misc updates

### Changes to actions `then()` callback.

Previously, if you was to call the action and use a callback in the
`then` method like so:

```php
(new CreateOrder)->execute(...)->then(function () {
  // ...
});
```

You wouldn't have access to the result of the action, which was a bit
pointless, now you have access to whatever the result was, like so:

```php
(new CreateOrder)->execute(...)->then(function ($order) {
  // ...
});
```

### Added `orderId` to the `PaymentAuthorize` DTO

Old:

```php
public function __construct(
    public bool $success = false,
    public ?string $message = null
) {
```

New:

```php
public function __construct(
    public bool $success = false,
    public ?string $message = null,
    public ?int $orderId = null
) {
```

Notice we don't actually pass the Order model, this is because orders
could potentially have many many lines which could cause problems when
serialised and by the time the data gets to the job queue etc it could
potentially be outdated, so it's up to your handlers to fetch the order
to ensure it's the most up to date version.


### Added `PaymentAttemptEvent`

There is now a new event, which will be used by add ons (like Stripe) to
dispatch when payment is attempted, emphasis on the word "attempted",
this will trigger whether successful or not and an instance of
`Lunar\Base\DataTransferObjects\PaymentAuthorize` will be passed through
to the event so you can check logic accordingly like so:

```php
final public function handle(PaymentAttemptEvent $event)
{
    $paymentAuthorize = $event->paymentAuthorize;

    if ($paymentAuthorize->success) {
        $order = Order::find($paymentAuthorize->orderId);

        Mail::to($order->billingAddress->contact_email)->queue(
            new OrderConfirmationMail($order)
        );
    }
}
```

---

The `order_id` column is deprecated on `carts` for ease of
migration/upgrade and figured it's better left to the next update to
remove entirely.