Discount query improvements (#1109)
Based on #1103 this PR looks to optimise the query when fetching
discounts. Beforehand we just loaded all available discounts, regardless
of whether they were actually needed and then left it up to the
individual types to figure out whether they should apply, this will be
quite taxing on sites with large amounts of discounts. This PR looks to
change the following:

- Removed the custom logic for checking customer groups / channels in
favour of the recently added scopes.
- Added `usable` scope to discounts so we can query based on whether
`uses` has exceeded `max_uses` if present.
- Added `products` scope to discounts to query based on whether the
discount is limited to certain product ids.
- The `Discounts::getDiscounts()` method now accepts a nullable `Cart`
model to be passed.
- If the cart exists, we only fetch discounts which have the products in
the cart is conditions or no conditions at all.
- If the cart has a coupon, we only fetch discounts which match that
coupon, or if the discount requires no coupon.

This PR relies on the tweak on #1108

---------

Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>