Installation process fix (#1493)
Just did a clean install of v1.0. And this step is in the wrong order. 

First enable the filament panel. Then do the `lunar:install`. Otherwise
an error occurs with the lunar:install, stating the panel is needed. But
then it's unclear that you need to run the installer again to seed the
database with needed permissions and users etc.

---------

Co-authored-by: Alec Ritson <hello@itsalec.co.uk>
Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>