Allow registration of the same extension on multiple pages (#1627)
The new `LunarPanel::extensions([])` registration method prevents the
reuse of the same extension across multiple pages as it will create
duplicate keys in the array.

As it's much more likely that an extension will be reused, rather than
having multiple extensions for a single page/resource, this update
switches the key/value on the array. This also aligns better with the
way the array is used in the Panel Manager.