Disable certain attribute fields when editing (#1691)
The hub doesn't really support the changing of certain attribute columns
once hey have been created. This is because it's hard to keep the
`attribute_data` in sync with what the new fieldtype might require,
leading to quite an overhead for something seemingly simple.

This PR will disable the `handle` and `type` columns when editing an
attribute to preserve and prevent data issues. If a developer needs to
change the fieldtype then it should be done in a migration or similar so
that appropriate steps can be taken to ensure the data is migrated
properly.

As a side effect this should close #1655

---------

Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>