Add soft deletes to carts (#1775)
This PR adds soft deletes to carts. The idea behind this is when you
want to `forget` a cart, it is likely you want to remove the cart from a
users current and future session.

The problem is in order to fully achieve this you would need to fully
remove the cart from the database, which means there is a loss of
important data as a store owner.

There are also situations you may want to forget the cart session, such
as logging out, but you may want to retain that cart for when the user
logs back in and can pick up from where they left off.

This PR looks to solve these issues by adding soft deletes to carts. The
default behaviour is the cart will be soft deleted when you call
`CartSession::forget`, but you can opt to keep the cart "active" by
passing the `delete: false` parameter.

### Existing Storefronts who want to soft delete carts
No further action is needed, call `CartSession::forget()` and the carts
should be deleted.

### Existing Storefronts who **do not** want to soft delete carts

```
// Old
CartSession::forget();

// New
CartSession::forget(delete: false);
```

---------

Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>