Fixes postgres compatibility on database state `PopulateProductOptionLabelWithName` updater (#1852)
fixes #1849 so that Lunar can be upgraded without issue using
PostgreSQL.

This has only been manually testing against Postgres 15.

The intent of the original version was, no doubt, to be more robust than
a simple null check. However the case of a non-valid, non-null value
existing in the product_options.label seems to be beyond the scope of a
simple script meant to backfill new columns with default data.

Implementing a check for an empty object or non-object value is tricky
without writting different checks for each DB engines.