Use correct base price in PricingResponse (#2080)
Base price shouldn't be customerGroup scoped, so use $basePrice where
customer_group_id is false.

The basePrice is initially retrieved correctly:

https://github.com/lunarphp/lunar/blob/467b49765d827ce937c7ce91043147195546ec59/packages/core/src/Managers/PricingManager.php#L198

then is not used in PricingResponse, instead is newly retrieved loosing
the condition that customer_group_id must be false.

https://github.com/lunarphp/lunar/blob/467b49765d827ce937c7ce91043147195546ec59/packages/core/src/Managers/PricingManager.php#L220

That makes potentially get the wrong price as basePrice