Added Price model method to return the tax inclusive compare price
Lunar provides a method to get the tax inclusive price but there's no method to get the compare price tax inclusive. If we're showing tax inclusive price while NOT storing prices as tax inclusive, then we will have a discrepancy between the compare price and price of product.
This commit adds a method to solve this issue.