Add ability to skip Lunar event forwarding for specific events
When extending Lunar models, events are fired for both the custom class and the Lunar base class. This can cause issues when you want to override event behavior in your custom model.

This change adds a $skipLunarEventForwarding property that allows developers to specify which events should NOT be forwarded to the Lunar base model.

Usage:

class Product extends LunarProduct
{
    protected array $skipLunarEventForwarding = ["created", "updated"];
}

This will fire the event only for the custom class, not for Lunar\Models\Product.